move picks function, pinpoint lines
