+++ mne/viz/utils.py
@@ -2496,18 +2496,6 @@ def _check_cov(noise_cov, info):
     return noise_cov
 
 
-def _get_channel_types(info, picks=None, unique=True,
-                       restrict_data_types=False):
-    """Get the data channel types in an info instance."""
-    picks = range(info['nchan']) if picks is None else picks
-    ch_types = [channel_type(info, idx) for idx in range(info['nchan'])
-                if idx in picks]
-    if restrict_data_types is True:
-        ch_types = [ch_type for ch_type in ch_types
-                    if ch_type not in _DATA_CH_TYPES_SPLIT]
-    return set(ch_types) if unique is True else ch_types
-
-
 def _set_title_multiple_electrodes(title, combine, ch_names, max_chans=6):
     """Prepare a title string for multiple electrodes."""
     title = ", ".join(ch_names[:max_chans]) if title is None else title
